refactor: replace old share by the new one
