+++ src/features/share/helpers/getShareOffer.tsx
@@ -1,7 +1,7 @@
 import { BookingOfferResponse, FavoriteOfferResponse, OfferResponse } from 'api/gen'
 import { formatShareOfferMessage } from 'features/share/helpers/formatShareOfferMessage'
 import { getOfferUrl } from 'features/share/helpers/getOfferUrl'
-import { share } from 'libs/share/shareBest'
+import { share } from 'libs/share/share'
 import { ShareContent } from 'libs/share/types'
 
 type Parameters = {
