--- src/libs/share/shareBest.web.test.tsx
@@ -1,17 +1,0 @@
-import { Linking } from 'react-native'
-
-import { share } from 'libs/share/shareBest'
-
-const mockOpenUrl = jest.spyOn(Linking, 'openURL').mockResolvedValue(undefined)
-
-const defaultContent = { body: 'Message', url: new URL('https://www.toto.com') }
-
-describe('share()', () => {
-  it('should share with whatsapp url if on web and whatsapp mode', () => {
-    share({ content: defaultContent, mode: 'WhatsApp' })
-
-    expect(mockOpenUrl).toHaveBeenCalledWith(
-      'https://api.whatsapp.com/send?text=Message%C2%A0%3A%0Ahttps%3A%2F%2Fwww.toto.com/'
-    )
-  })
-})
